refactor: move geostatistics tasks to tasks/geostatistics/ subpackage
Move kriging and break_ties task modules into a new
tasks/geostatistics/ directory to group tasks by their compute topic.

- Move kriging.py and break_ties.py to tasks/geostatistics/
- Create geostatistics/__init__.py with registration imports
- Update all relative imports (.common -> ..common)
- Update tasks/__init__.py to import from new location
- Update test imports across all test files
- Update notebook and documentation import paths
