+++ mkdocs/docs/packages/evo-compute/typed-objects/break-ties/BreakTiesParameters.md
@@ -1,0 +1,4 @@
+[GitHub source](https://github.com/SeequentEvo/evo-python-sdk/blob/main/packages/evo-compute/src/evo/compute/tasks/geostatistics/break_ties.py)
+::: evo.compute.tasks.geostatistics.break_ties.BreakTiesParameters
+    options:
+      show_labels: false
