+++ mkdocs/docs/packages/evo-compute/typed-objects/kriging/KrigingMethod.md
@@ -1,4 +1,4 @@
-[GitHub source](https://github.com/SeequentEvo/evo-python-sdk/blob/main/packages/evo-compute/src/evo/compute/tasks/kriging.py)
-::: evo.compute.tasks.kriging.KrigingMethod
+[GitHub source](https://github.com/SeequentEvo/evo-python-sdk/blob/main/packages/evo-compute/src/evo/compute/tasks/geostatistics/kriging.py)
+::: evo.compute.tasks.geostatistics.kriging.KrigingMethod
     options:
       show_labels: false
