+++ mkdocs/site/packages/evo-compute/index.html
@@ -86,7 +86,7 @@ JobClient provides low-level access to the API, while the <code>run()</code> fun
 <p>The <code>run()</code> function is the main entry point for executing compute tasks. It supports running a single task or multiple tasks concurrently.</p>
 <h3 id="single-task">Single task</h3>
 <pre><code class="language-python">from evo.compute.tasks import run, SearchNeighborhood, Ellipsoid, EllipsoidRanges
-from evo.compute.tasks.kriging import KrigingParameters
+from evo.compute.tasks.geostatistics.kriging import KrigingParameters
 
 params = KrigingParameters(
     source=pointset.attributes[&quot;grade&quot;],
@@ -102,7 +102,7 @@ result = await run(manager, params, preview=True)
 <h3 id="multiple-tasks">Multiple tasks</h3>
 <p>Run multiple kriging tasks concurrently — for example, estimating different attributes or using different parameters:</p>
 <pre><code class="language-python">from evo.compute.tasks import run, SearchNeighborhood
-from evo.compute.tasks.kriging import KrigingParameters
+from evo.compute.tasks.geostatistics.kriging import KrigingParameters
 
 results = await run(manager, [
     KrigingParameters(
@@ -145,7 +145,7 @@ df = await result.to_dataframe()
 <li>Run the <strong>remaining</strong> tasks — now <code>block_model.attributes["kriged_grade"]</code> resolves to the existing attribute and will update it.</li>
 </ol>
 <pre><code class="language-python">from evo.compute.tasks import run, SearchNeighborhood
-from evo.compute.tasks.kriging import KrigingParameters, RegionFilter
+from evo.compute.tasks.geostatistics.kriging import KrigingParameters, RegionFilter
 
 # Step 1: Run the first task — attribute &quot;kriged_grade&quot; does not exist yet, so it is created
 first_result = await run(manager, KrigingParameters(
