+++ mkdocs/typed_modules.txt
@@ -1,5 +1,5 @@
 packages.evo-objects.src.evo.objects.typed
 packages.evo-blockmodels.src.evo.blockmodels.typed
-packages.evo-compute.src.evo.compute.tasks
+packages.evo-compute.src.evo.compute.tasks.geostatistics
 packages.evo-compute.src.evo.compute.tasks.common
